fix compile errors and some warnings

+++ examples/upscale_avg.cpp
@@ -49,11 +49,7 @@
 #include <numeric> // for std::accumulate
 #include <sys/utsname.h>
 
-#include <dune/grid/common/EclipseGridParser.hpp>
-#include <dune/grid/common/EclipseGridInspector.hpp>
-//#include <dune/common/MonotCubicInterpolator.hpp>
 #include <dune/upscaling/SinglePhaseUpscaler.hpp>
-#include <dune/common/Units.hpp>
 
 using namespace Dune;
 using namespace std;
@@ -274,7 +270,7 @@ int main(int varnum, char** vararg) {
         
         int zeroporocells = 0;
         int negativeporocells = 0;
-        for (uint cell_idx = 0; cell_idx < num_eclipse_cells; ++cell_idx) {
+        for (uint cell_idx = 0; cell_idx < (uint)num_eclipse_cells; ++cell_idx) {
             if  (poros[cell_idx] == 0) {
                 ++zeroporocells;
             }
@@ -299,7 +295,7 @@ int main(int varnum, char** vararg) {
     if (doperm) {
         int zeropermcells = 0;
         int negativepermcells = 0;
-        for (uint cell_idx = 0; cell_idx < num_eclipse_cells; ++cell_idx) {
+        for (uint cell_idx = 0; cell_idx < (uint)num_eclipse_cells; ++cell_idx) {
             if (cellVolumes[cell_idx] > emptycellvolumecutoff) {
                 permxsum += permxs[cell_idx];
                 volpermxsum += permxs[cell_idx] * cellVolumes[cell_idx];
@@ -382,7 +378,7 @@ int main(int varnum, char** vararg) {
         int maxsatnumvalue = 0;
         // Check that SATNUM are set sensibly, that is > 0 and < 1000, and find number
         // of unique satnums present ( = number of rocktypes)
-        for (uint i = 0; i < satnums.size(); ++i) {
+        for (uint i = 0; i < (uint)satnums.size(); ++i) {
             if (satnums[i] > maxsatnumvalue) {
                 maxsatnumvalue = satnums[i];
             }
@@ -420,7 +416,7 @@ int main(int varnum, char** vararg) {
         activecellcount_rocktype.resize(maxsatnumvalue+1, 0);
 
         // Now loop over cells to collect statistics pr. rocktype
-        for (uint cell_idx = 0; cell_idx < num_eclipse_cells; ++cell_idx) {
+        for (uint cell_idx = 0; cell_idx < (uint)num_eclipse_cells; ++cell_idx) {
             ++totalcellcount_rocktype[satnums[cell_idx]];
             if (cellVolumes[cell_idx] > emptycellvolumecutoff) {
                 ++activecellcount_rocktype[satnums[cell_idx]];
@@ -445,7 +441,7 @@ int main(int varnum, char** vararg) {
           phi_var = 1/V*sum(v_i*(phi_i-phi_avg)^2)
         */
         if (doporosity) {
-            for (uint cell_idx = 0; cell_idx < num_eclipse_cells; ++cell_idx) {
+            for (uint cell_idx = 0; cell_idx < (uint)num_eclipse_cells; ++cell_idx) {
                 if (cellVolumes[cell_idx] > emptycellvolumecutoff) {
                     porosityvariancesum_rocktype[satnums[cell_idx]] += cellVolumes[cell_idx] 
                         * pow((poros[cell_idx]-porevolumesum_rocktype[satnums[cell_idx]]/volumesum_rocktype[satnums[cell_idx]]),2);
